add test, refactor for testability
