+++ package.json
@@ -29,6 +29,7 @@
   },
   "devDependencies": {
     "@babel/core": "~7.6.0",
+    "@testing-library/react-native": "^7.1.0",
     "@types/react": "~16.9.35",
     "@types/react-native": "~0.63.2",
     "babel-preset-expo": "^8.3.0",
