+++ src/components/ImportSiteTags/__tests__/ImportSiteTags.js
@@ -1,52 +1,67 @@
-import { parseSiteTagsAndOptions } from '../ImportSiteTags';
+import 'react-native';
+import React from 'react';
+import { render, fireEvent } from '@testing-library/react-native';
+import ImportSiteTags from '../ImportSiteTags';
+import parser from '../parser';
 
 
-describe('parseSiteTagsAndOptions', () => {
-  test('good input', () => {
-    const goodInput = `
-      <option value="dpmr8">foo</option>\n
-      <option value="g16">bar:8</option>\n
-      <option value="dpm22">baz:8080</option>\n
-    `;
-    const result = parseSiteTagsAndOptions(goodInput);
-    console.log(result);
-    expect(result).toMatchObject({
-      foo: {
-        newPasswordBumper: 0,
-        requireDigit: true,
-        requirePunctuation: false,
-        requireMixedCase: true,
-        noSpecial: true,
-        digitsOnly: false,
-        size: 8
-      },
-      bar: {
-        newPasswordBumper: 8,
-        requireDigit: true,
-        requirePunctuation: false,
-        requireMixedCase: false,
-        noSpecial: true,
-        digitsOnly: true,
-        size: 16,
-      },
-      'baz:8080': {
-        newPasswordBumper: 0,
-        requireDigit: true,
-        requirePunctuation: true,
-        requireMixedCase: true,
-        noSpecial: false,
-        digitsOnly: false,
-        size: 22,
-      },
+jest.mock('../parser');
+
+describe('ImportSiteTags', () => {
+  const noop = () => {};
+  const formattedSiteTags = `
+    <option value="dpm8">foo</option>
+  `;
+
+  describe('when parser throws', () => {
+    it('should display error and hide submit button', () => {
+      parser.parseSiteTagsAndOptions.mockImplementationOnce(() => {
+        throw 'mock error';
+      });
+      const { getByPlaceholderText, queryByTestId } = render(
+        <ImportSiteTags onCancel={noop} onSubmit={noop} siteTagList={[]} />
+      );
+      const textInput = getByPlaceholderText('Paste HTML');
+      fireEvent.changeText(textInput, formattedSiteTags);
+
+      expect(queryByTestId('error-container')).not.toBeNull();
+      expect(queryByTestId('submit-button')).toBeNull();
+    });
+  });
+
+  describe('when parser returns empty', () => {
+    it('should display error and hide submit button', () => {
+
+    });
+  });
+
+  describe('when parser returns only existing site tags', () => {
+    it('should display error and hide submit button', () => {
+
+    });
+  });
+
+  describe('when parser returns some new and some existing site tags', () => {
+    it('should display error and show submit button', () => {
+
+    });
+
+    describe('and user hits submit', () => {
+      it('should submit only new site tags', () => {
+
+      });
     })
   });
 
-  test('bad input', () => {
-    const badInput = `
-      n value="dpmr8">foo</op\ntion><optio
-      16">bar:8</oalue="g>\nption<option v
-      option>\ntion value="<opdpm22">baz:8080</
-    `;
-    expect(parseSiteTagsAndOptions(badInput)).toMatchObject({});
+  describe('when there are no issues', () => {
+    it('should display no error and show submit button', () => {
+
+    });
+
+    describe('and user hits submit', () => {
+      it('should pass site tag options', () => {
+
+      });
+    });
   });
 });
